Contacts and bugs

+++ client/app/scripts/controllers/contactController.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var myApp  = angular.module('sbAdminApp')
-myApp.controller('ContactCtrl', ['$scope' ,'$state','ContactService','$stateParams' , 'DTOptionsBuilder', 'DTColumnDefBuilder' ,function ($scope ,$state,ContactService,$stateParams, DTOptionsBuilder, DTColumnDefBuilder) {
+myApp.controller('ContactCtrl', ['$scope' ,'$state','ContactService','$stateParams' , 'DTOptionsBuilder', 'DTColumnDefBuilder' ,'$window' ,function ($scope ,$state,ContactService,$stateParams, DTOptionsBuilder, DTColumnDefBuilder , $window) {
 
 
 
@@ -26,20 +26,63 @@ myApp.controller('ContactCtrl', ['$scope' ,'$state','ContactService','$statePara
     })
 
 
-
+    $scope.modifyContact = function(contact,index) {
+        $state.go('dashboard.addContact', {contact : contact});
+    }
 
 
 
     $scope.removeContact = function(contact,index) {
-       
-        ContactService.delete(contact._id).then(function successCallback(response) {
-                
-                $scope.contacts.splice(index, 1);
-    
-        }, function errorCallback(response) {
 
-        })
+        var deleteContact = $window.confirm('Are you sure you want to delete ' +  contact.name + '?');
+        if(deleteContact){
+       
+            ContactService.delete(contact._id).then(function successCallback(response) {
+                    
+                    $scope.contacts.splice(index, 1);
+        
+            }, function errorCallback(response) {
+
+            })
+        }
     }
+
+}]);
+
+myApp.controller('AddContactCtrl', ['$scope' ,'$state','ContactService','$stateParams' , 'DTOptionsBuilder', 'DTColumnDefBuilder','ProductService' ,function ($scope ,$state,ContactService,$stateParams, DTOptionsBuilder, DTColumnDefBuilder,ProductService) {
+
+ init();
+
+
+
+ $scope.addContact = function(contact){
+    if(!contact.isEdit){
+     ContactService.create(contact)
+     .then(function successCallback(response) {
+        console.log(response);
+        $state.go('dashboard.show_contact',{contactId : response.data._id })
+
+    }, function errorCallback(response) {
+    })
+ }else{
+    ContactService.edit(contact).then(function successCallback(response) {
+
+        $state.go('dashboard.contact')
+
+    }, function errorCallback(response) {
+
+    })
+}
+}
+
+
+function init(){
+    if($stateParams && $stateParams.contact){
+     $scope.contact = $stateParams.contact;
+     $scope.contact.isEdit = true;            
+ }
+}
+
 
 }]);
